Fix crash

+++ WowPacketParser/Store/Objects/Unit.cs
@@ -36,7 +36,7 @@ namespace WowPacketParser.Store.Objects
         public float? Size;
         public uint? Bytes0;
         public uint? MaxHealth;
-        public uint? Level;
+        public int? Level;
         public uint? Faction;
         public uint[] Equipment;
         public UnitFlags? UnitFlags;
@@ -95,7 +95,7 @@ namespace WowPacketParser.Store.Objects
             Bytes2        = UpdateFields.GetValue<UnitField, uint?>(UnitField.UNIT_FIELD_BYTES_2);
             Size          = UpdateFields.GetValue<ObjectField, float?>(ObjectField.OBJECT_FIELD_SCALE_X);
             MaxHealth     = UpdateFields.GetValue<UnitField, uint?>(UnitField.UNIT_FIELD_MAXHEALTH);
-            Level         = UpdateFields.GetValue<UnitField, uint?>(UnitField.UNIT_FIELD_LEVEL);
+            Level         = (int)UpdateFields.GetValue<UnitField, uint?>(UnitField.UNIT_FIELD_LEVEL);
             Faction       = UpdateFields.GetValue<UnitField, uint?>(UnitField.UNIT_FIELD_FACTIONTEMPLATE);
             Equipment     = UpdateFields.GetArray<UnitField, uint>(UnitField.UNIT_VIRTUAL_ITEM_SLOT_ID1, 3);
             UnitFlags     = UpdateFields.GetEnum<UnitField, UnitFlags?>(UnitField.UNIT_FIELD_FLAGS);
